chore: MVC 구조를 위한 view, model 분리 및 사용자에게 보일 입력출값을 위한 함수합침 수정
