+++ src/model/Racing.js
@@ -3,9 +3,11 @@ import { Random } from "@woowacourse/mission-utils";
 class Racing{
 
     #scoreNumber
+    #carName
 
     constructor(racingCar){
         this.#scoreNumber = [];
+        this.#carName = [];
         this.racing = racingCar;
     }
 
@@ -33,6 +35,13 @@ class Racing{
         const MAX_VALUE = Math.max(...this.#scoreNumber);
         const RACING_MAX_VALUE = this.#racingScoreCompare(MAX_VALUE);
         return RACING_MAX_VALUE
+    }
+
+    carNameUpdate = (carName) => {
+        return carName.reduce((arr, cur, idx) => {
+            this.#carName.push(cur+' : ');
+            return this.#carName;
+        }, "");
     }
 
     #racingScoreCompare = (maxValue) => {
